Fixed warnings as output by clang for 64-bit Intel compilation as given when using Xcode 7.1
* Most of them are of the 64-bit -> 32-bit implicit conversion, and they have been patched by using static_cast (or C cast in C files). Many of them are resulted by getting the item index from an array.
* Printf format warnings have been patched by casting the argument, while keeping the format string as-is
It should be up to the main developer to decide what to do with all these casts. Hopefully using static_cast explicitly will aid in finding all these cases.

* A few others are about missing variables. Unless they were already marked with FIXME or TODO, I just marked them too, instead of deleting or commenting them out.

+++ source/acs_acs0.cpp
@@ -696,7 +696,7 @@ static void ACS_translateScriptACS0(acs0_tracer_t *tracer, int32_t *codeIndexMap
       opdata = &ACS0opdata[op];
 
       // Record jump target.
-      codeIndexMap[index] = codePtr - tracer->vm->code;
+      codeIndexMap[index] = static_cast<int32_t>(codePtr - tracer->vm->code);
 
       // Calculate next index.
       index += ACS_countOpSizeACS0(tracer, index, opSize, opdata);
@@ -1250,7 +1250,7 @@ uint32_t ACS_LoadStringACS0(const byte *begin, const byte *end)
          buf += *itr++;
    }
 
-   return ACSVM::AddString(buf.constPtr(), buf.length());
+   return ACSVM::AddString(buf.constPtr(), static_cast<uint32_t>(buf.length()));
 }
 
 // EOF
