+++ source/g_game.cpp
@@ -1337,7 +1337,7 @@ static void G_ReadDemoTiccmd(ticcmd_t *cmd)
 //
 static void G_WriteDemoTiccmd(ticcmd_t *cmd)
 {
-   unsigned int position = demo_p - demobuffer;
+   unsigned int position = static_cast<unsigned>(demo_p - demobuffer);
    int i = 0;
    
    demo_p[i++] = cmd->forwardmove;
@@ -2368,7 +2368,7 @@ static bool G_CheckSpot(int playernum, mapthing_t *mthing, Mobj **fog)
 //
 int G_ClosestDMSpot(fixed_t x, fixed_t y, int notspot)
 {
-   int j, numspots = deathmatch_p - deathmatchstarts;
+   int j, numspots = static_cast<int>(deathmatch_p - deathmatchstarts);
    int closestspot = -1;
    fixed_t closestdist = 32767*FRACUNIT;
 
@@ -2400,7 +2400,7 @@ extern const char *level_error;
 //
 void G_DeathMatchSpawnPlayer(int playernum)
 {
-   int j, selections = deathmatch_p - deathmatchstarts;
+   int j, selections = static_cast<int>(deathmatch_p - deathmatchstarts);
    Mobj *fog = NULL;
    
    if(selections < MAXPLAYERS)
