+++ source/p_map3d.cpp
@@ -991,7 +991,7 @@ static bool midtex_moving;
 //
 static int P_PushUp(Mobj *thing)
 {
-   unsigned int firstintersect = intersectors.getLength();
+   unsigned int firstintersect = static_cast<unsigned>(intersectors.getLength());
    unsigned int lastintersect;
    int mymass = thing->info->mass;
 
@@ -999,7 +999,7 @@ static int P_PushUp(Mobj *thing)
       return 1;
 
    P_FindAboveIntersectors(thing);
-   lastintersect = intersectors.getLength();
+   lastintersect = static_cast<unsigned>(intersectors.getLength());
    for(; firstintersect < lastintersect; ++firstintersect)
    {
       Mobj *intersect = intersectors[firstintersect];
@@ -1035,7 +1035,7 @@ static int P_PushUp(Mobj *thing)
 //
 static int P_PushDown(Mobj *thing)
 {
-   unsigned int firstintersect = intersectors.getLength();
+   unsigned int firstintersect = static_cast<unsigned>(intersectors.getLength());
    unsigned int lastintersect;
    int mymass = thing->info->mass;
 
@@ -1043,7 +1043,7 @@ static int P_PushDown(Mobj *thing)
       return 1;
 
    P_FindBelowIntersectors(thing);
-   lastintersect = intersectors.getLength();
+   lastintersect = static_cast<unsigned>(intersectors.getLength());
    for(; firstintersect < lastintersect; ++firstintersect)
    {
       Mobj *intersect = intersectors[firstintersect];
